+++ src/Krystal/Authentication/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 1.3
 ---
 
+ * Added `storeData()` and `getData()` methods in `authService`
  * Fixed issue with throwing exception for non-login attempts
  * Added protection service for failure attempts